Prevent croaking when JIRA uses a %F-type date
I came across an issue in our JIRA project at work that had a duedate:
'2017-08-01'. Turning that issue into an object made
JIRA::REST::Class::Factory::make_date() croak because it wasn't prepared for a
date without a timestamp.
This patch makes make_date() have a look at the date first to decide on the
correkt pattern to use for parsing the date.